Replacing server with server_v2. Updating docs
