+++ src/deepsparse/server/cli.py
@@ -12,6 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+"""
+There are two sub-commands for the server:
+1. `deepsparse.server config [OPTIONS] <config path>`
+2. `deepsparse.server quick [OPTIONS] <task>
+```
+"""
+
 import os
 from tempfile import TemporaryDirectory
 from typing import Optional
@@ -20,12 +27,77 @@ import click
 import yaml
 
 from deepsparse.pipeline import Pipeline, SupportedTasks
-from deepsparse.server_v2.config import EndpointConfig, ServerConfig
-from deepsparse.server_v2.main import start_server
+from deepsparse.server.config import EndpointConfig, ServerConfig
+from deepsparse.server.config_server import start_server
 
 
-@click.command(
-    context_settings=dict(token_normalize_func=lambda x: x.replace("-", "_"))
+@click.group()
+def main():
+    """
+    Start a DeepSparse inference server for serving the models and pipelines.
+
+        1. `deepsparse.server config [OPTIONS] <config path>`
+
+        2. `deepsparse.server task [OPTIONS] <task>
+
+    Examples for using the server:
+
+        `deepsparse.server config server-config.yaml`
+
+        `deepsparse.server task question_answering --batch-size 2`
+
+        `deepsparse.server task question_answering --host "0.0.0.0"`
+
+    Example config.yaml for serving:
+
+    ```yaml
+    num_cores: 2
+    num_workers: 2
+    endpoints:
+        - task: question_answering
+          endpoint: /unpruned/predict
+          model_path: zoo:some/zoo/stub
+        - task: question_answering
+          endpoint: /pruned/predict
+          model_path: /path/to/local/model
+    ```
+    """
+    pass
+
+
+@main.command()
+@click.argument("config-path", type=str)
+@click.option(
+    "--host",
+    type=str,
+    default="0.0.0.0",
+    help=(
+        "Bind socket to this host. Use --host 0.0.0.0 to make the application "
+        "available on your local network. "
+        "IPv6 addresses are supported, for example: --host '::'. Defaults to 0.0.0.0"
+    ),
+)
+@click.option(
+    "--port",
+    type=int,
+    default=5543,
+    help="Bind to a socket with this port. Defaults to 5543.",
+)
+@click.option(
+    "--log-level",
+    type=click.Choice(
+        ["debug", "info", "warn", "critical", "fatal"], case_sensitive=False
+    ),
+    default="info",
+    help="Sets the logging level. Defaults to info.",
+)
+def config(config_path: str, host: str, port: int, log_level: str):
+    "Run the server using configuration from a .yaml file."
+    start_server(config_path, host, port, log_level)
+
+
+@main.command(
+    context_settings=dict(token_normalize_func=lambda x: x.replace("-", "_")),
 )
 @click.argument(
     "task",
@@ -35,9 +107,11 @@ from deepsparse.server_v2.main import start_server
     "--model_path",
     type=str,
     default=None,
-    help="The path to a model.onnx file, a model folder containing the model.onnx "
-    "and supporting files, or a SparseZoo model stub. "
-    "Ignored if config_file is supplied.",
+    help=(
+        "The path to a model.onnx file, a model folder containing the model.onnx "
+        "and supporting files, or a SparseZoo model stub. "
+        "If not specified, the default model for the task is used."
+    ),
 )
 @click.option(
     "--batch_size",
@@ -74,14 +148,14 @@ from deepsparse.server_v2.main import start_server
     help="Bind to a socket with this port. Defaults to 5543.",
 )
 @click.option(
-    "--log_level",
+    "--log-level",
     type=click.Choice(
         ["debug", "info", "warn", "critical", "fatal"], case_sensitive=False
     ),
     default="info",
     help="Sets the logging level. Defaults to info.",
 )
-def main(
+def task(
     task: str,
     model_path: Optional[str],
     batch_size: int,
@@ -91,6 +165,10 @@ def main(
     port: int,
     log_level: str,
 ):
+    """
+    Run the server using configuration with CLI options,
+    which can only server a single model.
+    """
     cfg = ServerConfig(
         num_cores=num_cores,
         num_workers=num_workers,
