--- src/deepsparse/server_v2/config.py
@@ -1,162 +1,0 @@
-# Copyright (c) 2021 - present / Neuralmagic, Inc. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-from typing import Any, Dict, List, Optional, Tuple, Union
-
-from pydantic import BaseModel, Field
-
-from deepsparse import DEEPSPARSE_ENGINE, PipelineConfig
-from deepsparse.tasks import SupportedTasks
-
-
-__all__ = [
-    "ServerConfig",
-    "EndpointConfig",
-    "SequenceLengthsConfig",
-    "ImageSizesConfig",
-]
-
-
-class SequenceLengthsConfig(BaseModel):
-    sequence_lengths: List[int] = Field(
-        description="The sequence lengths the model should accept"
-    )
-
-
-class ImageSizesConfig(BaseModel):
-    image_sizes: List[Tuple[int, int]] = Field(
-        description="The list of image sizes the model should accept"
-    )
-
-
-class EndpointConfig(BaseModel):
-    name: Optional[str] = Field(
-        default=None,
-        description=(
-            "Name of the model used for logging & metric purposes. "
-            "If not specified 'endpoint-<index>' will be used."
-        ),
-    )
-
-    endpoint: Optional[str] = Field(
-        default=None,
-        description="Optional url to use for this endpoint. E.g. '/predict'. "
-        "If not specified '/endpoint-<index>/predict' will be used",
-    )
-
-    task: str = Field(description="Task this endpoint performers")
-
-    model: str = Field(description="Location of the underlying model to use.")
-
-    batch_size: int = Field(
-        default=1, description="The batch size to compile the model for."
-    )
-
-    bucketing: Optional[Union[ImageSizesConfig, SequenceLengthsConfig]] = Field(
-        default=None,
-        description=(
-            "What input shapes this model can accept."
-            "Example for multiple sequence lengths in yaml: "
-            "```yaml\n"
-            "bucketing:\n"
-            "  sequence_lengths: [16, 32, 64]\n"
-            "```\n"
-        ),
-    )
-
-    def to_pipeline_config(self) -> PipelineConfig:
-        input_shapes, kwargs = _unpack_bucketing(self.task, self.bucketing)
-
-        return PipelineConfig(
-            task=self.task,
-            model_path=self.model,
-            engine_type=DEEPSPARSE_ENGINE,
-            batch_size=self.batch_size,
-            num_cores=None,  # this will be set from Context
-            input_shapes=input_shapes,
-            kwargs=kwargs,
-        )
-
-
-class ServerConfig(BaseModel):
-    num_cores: int = Field(
-        description="The number of cores available for model execution. "
-        "Defaults to all available cores.",
-    )
-
-    num_workers: int = Field(
-        description="The number of workers to split the available cores between."
-    )
-
-    endpoints: List[EndpointConfig] = Field(description="The models to serve.")
-
-
-def _unpack_bucketing(
-    task: str, bucketing: Optional[Union[SequenceLengthsConfig, ImageSizesConfig]]
-) -> Tuple[Optional[List[int]], Dict[str, Any]]:
-    """
-    :return: (input_shapes, kwargs) which are passed to PipelineConfig
-    """
-    if bucketing is None:
-        return None, {}
-
-    if isinstance(bucketing, SequenceLengthsConfig):
-        if not SupportedTasks.is_nlp(task):
-            raise ValueError(f"SequenceLengthConfig specified for non-nlp task {task}")
-
-        return _unpack_nlp_bucketing(bucketing)
-    elif isinstance(bucketing, ImageSizesConfig):
-        if (
-            not SupportedTasks.is_image_classification(task)
-            and not SupportedTasks.is_yolo(task)
-            and not SupportedTasks.is_yolact(task)
-        ):
-            raise ValueError(
-                f"ImageSizeConfig specified for non computer vision task {task}"
-            )
-
-        return _unpack_cv_bucketing(bucketing)
-    else:
-        raise ValueError(f"Unknown bucket config {bucketing}")
-
-
-def _unpack_nlp_bucketing(cfg: SequenceLengthsConfig):
-    if len(cfg.sequence_lengths) == 0:
-        raise ValueError("Must specify at least one sequence length under bucketing")
-
-    if len(cfg.sequence_lengths) == 1:
-        input_shapes = None
-        kwargs = {"sequence_length": cfg.sequence_lengths[0]}
-    else:
-        input_shapes = None
-        kwargs = {"sequence_length": cfg.sequence_lengths}
-
-    return input_shapes, kwargs
-
-
-def _unpack_cv_bucketing(cfg: ImageSizesConfig):
-    if len(cfg.image_sizes) == 0:
-        raise ValueError("Must specify at least one image size under bucketing")
-
-    if len(cfg.image_sizes) == 1:
-        # NOTE: convert from List[Tuple[int, int]] to List[List[int]]
-        input_shapes = [list(cfg.image_sizes[0])]
-        kwargs = {}
-    else:
-        raise NotImplementedError(
-            "Multiple image size buckets is currently unsupported"
-        )
-
-    return input_shapes, kwargs
